+++ packages/components/modules/content-feed/common/index.ts
@@ -1,0 +1,3 @@
+// exports common content-feed code
+
+export * from './graphql/mutations/ContentPostCreate'