+++ packages/components/modules/content-feed/web/ContentFeedImage/types.ts
@@ -1,0 +1,12 @@
+import { UseFormReturn } from 'react-hook-form'
+
+export interface IContentFeedImageProps {
+  form: UseFormReturn<
+    {
+      content: string
+      images: File[]
+    },
+    any,
+    undefined
+  >
+}