+++ packages/components/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @baseapp-frontend/components
 
+## 1.0.34
+
+### Patch Changes
+- Content Feed creation page and Dropzone Improvements
+- Updated dependencies
+  - @baseapp-frontend/design-system@1.0.15
+
 ## 1.0.33
 
 ### Patch Changes